+++ packages/scripts/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## 1.4.2 (unreleased)
+## 1.5.0 (2022-07-28)
 
 ### Feature